% Generated by roxygen2 (4.0.1): do not edit by hand
\name{del}
\alias{del}
\title{Delete Columns from a data.table}
\usage{
del(DT, ..., list = NULL)
}
\arguments{
\item{DT}{A \code{data.table} object.}

\item{...}{A set of (quoted or unquoted) variable names.}

\item{list}{Optional: a list, or vector, of variable names from which to
search \code{DT} for variables to delete.}
}
\description{
This function deletes columns from a \code{data.table} through \code{set}.
This modifies the object in-place, and hence you do not need to re-assign
the result.
}
\examples{
local({
DT <- data.table( x=1:5, y=letters[1:5], z=factor( letters[1:5] ) )
str(DT)
del(DT, x)
str(DT)
vars_to_remove <- c("y", "q")
del(DT, list=vars_to_remove)
str(DT)
})
}

