% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues.R
\name{mapvalues}
\alias{mapvalues}
\title{Replace values of a vector based on a mapping scheme.}
\usage{
mapvalues(x, from, to)
}
\arguments{
\item{x}{A vector.}

\item{from}{A vector specifying the values that should be mapped.}

\item{to}{A vector specifying the values to map to.}
}
\value{
A vector. It may be of a different type than the input vector \code{x},
   depending on your inputs for \code{from} and \code{to}.
}
\description{
This is basically a copy or reimplementation of the
   \code{plyr::mapvalues()} function.
   It takes an input vector \code{x} as well as two vectors \code{from} and
   \code{to} that act as a dictionary for the mapping.
}
