% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swap}
\alias{swap}
\title{Swap Elements in a Vector}
\usage{
swap(vec, from, to = names(from), ...)
}
\arguments{
\item{vec}{the vector of items whose elements you will be replacing.}

\item{from}{the items you will be mapping 'from'.}

\item{to}{the items you will be mapping 'to'. must be same length and
order as \code{from}.}

\item{...}{optional arguments passed to \code{match}.}
}
\description{
This function swaps elements in a vector. See examples for usage.
}
\seealso{
\code{\link{match}}
}

