\name{rbindlistn}
\alias{rbindlistn}
\title{rbindlist with names}
\usage{
  rbindlistn(l, names = ".Names")
}
\arguments{
  \item{l}{A list of \code{data.frame}s, \code{data.table}s
  or \code{data.frame} shaped \code{list}s.}

  \item{names}{If \code{names} is: \itemize{
  \item{\code{FALSE}: Do not generate a vector of names.}
  \item{\code{TRUE}: Generate a vector of names called
  \code{.Names}, and append it to the \code{data.table}
  generated by \code{rbindlist}.} \item{A string: Generate
  a vector of names called \code{names}, and append it to
  the \code{data.table} generated by \code{rbindlist}.} }}
}
\description{
  This function implements \code{rbindlist} but adds an
  optional argument, \code{names}, which allows us to keep
  the names of each sub-list which we \code{rbindlist}-ify.
}
\examples{
dfs <- list( a=data.frame(x=1, y=2, z=3), b=data.frame(x=4, y=5, z=6) )
rbindlistn(dfs)
lists <- list( a=list(x=1, y=2, z=3), b=list(x=4, y=5, z=6) )
rbindlistn(lists)
}

