% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setj.R
\name{setj}
\alias{setj}
\alias{setj_at}
\alias{setj_if}
\alias{setj_grep}
\title{Conditional column manipulation by reference}
\usage{
setj_at(DT, cols, .f, ...)

setj_if(DT, .p, .f, ...)

setj_grep(DT, pattern, .f, ...)
}
\description{
These functions manipulate existing columns of your \code{data.table} by reference
based on a selection of columns, a predicate function or a regex pattern that the
column name must match.
}
